% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\title{Checagen de valores missing}
\usage{
check_missing(data)
}
\arguments{
\item{data}{Nome do dataframe}
}
\value{
Retorna uma tabela com o nome da coluna e a porcentagem de valores vazios
}
\description{
Checagen de valores missing
}
