% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_1.R
\name{tab_1}
\alias{tab_1}
\title{Tabela de frequência para uma variável}
\usage{
tab_1(df, coluna)
}
\arguments{
\item{df}{Nome do seu DataFrame}

\item{coluna}{Nome da coluna que deseja fazer a estatística}
}
\value{
Retorna uma tabela com a frequência e a porcentagem
}
\description{
Tabela de frequência para uma variável
}
