% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_2.R
\name{tab_2}
\alias{tab_2}
\title{Tabela cruzada de frequência para duas variáveis}
\usage{
tab_2(df, var_row, var_col, pct = FALSE, pct_row = FALSE)
}
\arguments{
\item{df}{Nome do DataFrame}

\item{var_row}{Variável que deseja ver como linha}

\item{var_col}{Variáveis que virarão colunas}

\item{pct}{Retorna frequência ou porcentagem. Por default, a tabela retorna a contagem das categorias, caso use}

\item{pct_row}{Retorna a frequência ou porcentagem com o total pela linha
pct = TRUE, a tabela retornará como porcentagem contabilizando 100\% por coluna}
}
\value{
Retorna uma tabela cruzando duas variáveis distintas, pode ser número absoluto (pct=FALSE), ou a porcentagem
pela coluna (pct=TRUE)
}
\description{
Tabela cruzada de frequência para duas variáveis
}
