% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_cat_sinan.R
\name{tab_cat_sinan}
\alias{tab_cat_sinan}
\title{Tabela para análise de colunas no SINAN violências}
\usage{
tab_cat_sinan(df, list, col, pct_reg = FALSE, pct_mul = FALSE)
}
\arguments{
\item{df}{Nome do dataframe}

\item{list}{Escolha da lista que será utilizada como categoria, ex: viol, enc, rel}

\item{col}{Valores nas colunas, ex: ds_raca, faixa_etaria_padrao}

\item{pct_reg}{Selecione se quer porcentagem por registro}

\item{pct_mul}{Selecione se quer porcentagem por número de mulheres}
}
\value{
Retorna tabela com as variáveis escolhidas na linha cruzada pela variável escolhida para a coluna
}
\description{
Tabela para análise de colunas no SINAN violências
}
