% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_multi.R
\name{tab_multi}
\alias{tab_multi}
\title{Tabela pra varias colunas com categorias iguais}
\usage{
tab_multi(df, lista, filtro)
}
\arguments{
\item{df}{Nome do DataFrame}

\item{lista}{Nome da lista com o nome das colunas}

\item{filtro}{Nome do banco que contém as colunas}
}
\value{
Dataframe em que as linhas são os nomes das colunas, as colunas são as frequências de categorias = 1 no banco
}
\description{
Tabela pra varias colunas com categorias iguais
}
