% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_bonita.R
\name{tabela_bonita}
\alias{tabela_bonita}
\title{Padronizar design de tabela}
\usage{
tabela_bonita(df, titulo, text_row = FALSE)
}
\arguments{
\item{df}{Tabela que receberá o design da tabela}

\item{titulo}{Título da tabela}

\item{text_row}{Transforma os textos da linha com letras minúsculas}
}
\value{
Retorna o Dataframe no fundo de tabela estética
}
\description{
Padronizar design de tabela
}
